extracted out case data into separate files
